requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-ai==0.0.12
google-cloud-vision==3.4.4
python-dotenv==1.0.0
google-generativeai==0.3.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
